requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
factory-boy==3.0.1
Faker==4.1.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
django-money==1.1

